split chunk of code out of main_enclone.rs
